Created a new Silver Surfer with Guess Factor Targeting
So I made this brand new robot based on the original Silver Surfer but this one has a guess factor gun.
The problem is it doesn't actually shoot at nearly the right angle for some reason. The onPaint() method
seems to draw he line of where it SHOULD shoot properly, but it doesn't actually shoot along that line.